remify: Processing and Transforming Relational Event History Data

Efficiently processes relational event history data and transforms them into formats suitable for other packages. The primary objective of this package is to convert event history data into a format that integrates with the packages in 'remverse' and is compatible with various analytical tools (e.g., computing network statistics, estimating tie-oriented or actor-oriented social network models). Second, it can also transform the data into formats compatible with other packages out of 'remverse'. The package processes the data for two types of temporal social network models: tie-oriented modeling framework (Butts, C., 2008, doi:10.1111/j.1467-9531.2008.00203.x ) and actor-oriented modeling framework (Stadtfeld, C., & Block, P., 2017, doi:10.15195/v4.a14 ).

See also

Author

Maintainer: Giuseppe Arena g.arena@tilburguniversity.edu (ORCID)

Other contributors:

Rumana Lakdawala r.j.lakdawala@tilburguniversity.edu [contributor]

Marlyne Meijerink-Bosman [contributor]

Diana Karimova d.karimova@tilburguniversity.edu [contributor]

Fabio Generoso Vieira f.v.generosovieira@tilburguniversity.edu [contributor]

Mahdi Shafiee Kamalabad m.shafieekamalabad@uu.nl [contributor]

Roger Leenders r.t.a.j.leenders@tilburguniversity.edu [contributor]

Joris Mulder j.mulder3@tilburguniversity.edu [contributor]